requirements.txt
matplotlib~=3.5.3
pillow~=11.0.0
openpyxl~=3.1.5
opencv-python~=4.8.0.76
numpy~=1.26.4
astropy~=6.0.1
scipy~=1.13.1
scikit-image~=0.24.0
scikit-learn~=1.6.0